# copilot-network-analyzer/connector/apiDefinition.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ntopng Network Analyzer",
    "description": "Read-only network visibility over ntopng: active hosts, live flows, protocol breakdown and alerts. Built to back a Microsoft Copilot Studio agent, but usable by any HTTP client.",
    "version": "0.1.0"
  },
  "host": "REPLACE-WITH-YOUR-PUBLIC-HOSTNAME",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/interfaces": {
      "get": {
        "operationId": "ListInterfaces",
        "summary": "List the interfaces ntopng is monitoring",
        "description": "Returns the interface ids that other operations accept. Most deployments have exactly one.",
        "x-ms-visibility": "important",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/InterfaceList"
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "502": {
            "description": "ntopng returned an error."
          },
          "503": {
            "description": "ntopng is unreachable."
          }
        }
      }
    },
    "/v1/summary": {
      "get": {
        "operationId": "GetNetworkSummary",
        "summary": "Get an overall snapshot of the network right now",
        "description": "One call returning active host and flow counts, aggregate throughput, the busiest hosts and the heaviest protocols. Prefer this for broad, opening questions such as 'how is my network doing', 'anything unusual', or 'give me a status update' \u2014 it avoids several narrower calls.",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "top",
            "in": "query",
            "required": false,
            "description": "How many top talkers and protocols to include.",
            "x-ms-summary": "Top",
            "type": "integer",
            "default": 5,
            "minimum": 1,
            "maximum": 20
          },
          {
            "name": "interface_id",
            "in": "query",
            "required": false,
            "description": "ntopng interface id. Defaults to the configured one.",
            "x-ms-summary": "Interface Id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/NetworkSummary"
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "502": {
            "description": "ntopng returned an error."
          },
          "503": {
            "description": "ntopng is unreachable."
          }
        }
      }
    },
    "/v1/alerts/recent": {
      "get": {
        "operationId": "GetRecentAlerts",
        "summary": "List recent ntopng alerts",
        "description": "Returns alerts ntopng raised over a recent window. Use this for 'anything wrong', 'any security alerts', or 'what happened overnight'. Some ntopng Community builds do not expose the alerts API \u2014 the response then has supported=false, and the agent should say alerting is unavailable rather than that the network is clean.",
        "x-ms-visibility": "important",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "description": "How far back to look, in hours.",
            "x-ms-summary": "Hours",
            "type": "integer",
            "default": 24,
            "minimum": 1,
            "maximum": 168
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "How many alerts to return.",
            "x-ms-summary": "Limit",
            "type": "integer",
            "default": 20,
            "minimum": 1,
            "maximum": 100
          },
          {
            "name": "interface_id",
            "in": "query",
            "required": false,
            "description": "ntopng interface id. Defaults to the configured one.",
            "x-ms-summary": "Interface Id",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "schema": {
              "$ref": "#/definitions/AlertList"
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "502": {
            "description": "ntopng returned an error."
          },
          "503": {
            "description": "ntopng is unreachable."
          }
        }
      }
    },
...
  },
  "definitions": {
    "Alert": {
      "properties": {
        "time": {
          "type": "integer",
          "title": "Time",
          "description": "Unix timestamp when the alert fired.",
          "x-nullable": true
        },
        "severity": {
          "type": "string",
          "title": "Severity",
          "description": "Alert severity, e.g. info, warning, error.",
          "x-nullable": true
        },
        "alert_type": {
          "type": "string",
          "title": "Alert Type",
          "description": "Alert type as named by ntopng.",
          "x-nullable": true
        },
        "entity": {
          "type": "string",
          "title": "Entity",
          "description": "Host or entity the alert is about.",
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "title": "Message",
          "description": "Human-readable alert text.",
          "x-nullable": true
        }
      },
      "type": "object",
      "title": "Alert"
    },
    "AlertList": {
      "properties": {
        "summary": {
          "type": "string",
          "title": "Summary",
          "description": "One-sentence, human-readable summary of the result."
        },
        "supported": {
          "type": "boolean",
          "title": "Supported",
          "description": "False when this ntopng build does not expose the alerts API."
        },
        "count": {
          "type": "integer",
          "title": "Count",
          "description": "Number of alerts returned."
        },
        "alerts": {
          "items": {
            "$ref": "#/definitions/Alert"
          },
          "type": "array",
          "title": "Alerts",
          "description": "Recent alerts, newest first."
        }
      },
      "type": "object",
      "required": [
        "summary",
        "supported",
        "count",
        "alerts"
      ],
      "title": "AlertList"
    },
...
    "Host": {
      "properties": {
        "ip": {
          "type": "string",
          "title": "Ip",
          "description": "IP address of the host."
        },
        "name": {
          "type": "string",
          "title": "Name",
          "description": "Resolved name (mDNS/DHCP/reverse DNS), or the IP when unresolved."
        },
        "is_local": {
          "type": "boolean",
          "title": "Is Local",
          "description": "True when the host is inside a configured local network."
        },
        "country": {
          "type": "string",
          "title": "Country",
          "description": "Two-letter country code for remote hosts.",
          "x-nullable": true
        },
        "throughput_bps": {
          "type": "number",
          "title": "Throughput Bps",
          "description": "Current throughput in bits per second."
        },
        "bytes_total": {
          "type": "integer",
          "title": "Bytes Total",
          "description": "Total bytes seen for this host since ntopng started tracking it."
        },
        "bytes_sent": {
          "type": "integer",
          "title": "Bytes Sent",
          "description": "Bytes sent by the host."
        },
        "bytes_received": {
          "type": "integer",
          "title": "Bytes Received",
          "description": "Bytes received by the host."
        },
        "active_flows": {
          "type": "integer",
          "title": "Active Flows",
          "description": "Number of currently active flows involving the host."
        },
        "last_seen": {
          "type": "integer",
          "title": "Last Seen",
          "description": "Unix timestamp of the last packet seen.",
          "x-nullable": true
        }
      },
      "type": "object",
      "required": [
        "ip",
        "name",
        "is_local",
        "throughput_bps",
        "bytes_total",
        "bytes_sent",
        "bytes_received",
        "active_flows"
      ],
      "title": "Host"
    },
...
    "Interface": {
      "properties": {
        "id": {
          "type": "integer",
          "title": "Id",
          "description": "ntopng interface id, passed as interface_id to other operations."
        },
        "name": {
          "type": "string",
          "title": "Name",
          "description": "Interface name as configured in ntopng."
        }
      },
      "type": "object",
      "required": [
        "id",
        "name"
      ],
      "title": "Interface"
    },
    "InterfaceList": {
      "properties": {
        "summary": {
          "type": "string",
          "title": "Summary",
          "description": "One-sentence, human-readable summary of the result."
        },
        "interfaces": {
          "items": {
            "$ref": "#/definitions/Interface"
          },
          "type": "array",
          "title": "Interfaces",
          "description": "Interfaces ntopng is monitoring."
        }
      },
      "type": "object",
      "required": [
        "summary",
        "interfaces"
      ],
      "title": "InterfaceList"
    },
    "NetworkSummary": {
      "properties": {
        "summary": {
          "type": "string",
          "title": "Summary",
          "description": "One-paragraph, human-readable state of the network. Safe to quote verbatim."
        },
        "interface_name": {
          "type": "string",
          "title": "Interface Name",
          "description": "Name of the monitored ntopng interface.",
          "x-nullable": true
        },
        "active_hosts": {
          "type": "integer",
          "title": "Active Hosts",
          "description": "Number of hosts currently active."
        },
        "local_hosts": {
          "type": "integer",
          "title": "Local Hosts",
          "description": "Number of active hosts inside the local networks."
        },
        "active_flows": {
          "type": "integer",
          "title": "Active Flows",
          "description": "Number of currently active flows."
        },
        "throughput_bps": {
          "type": "number",
          "title": "Throughput Bps",
          "description": "Aggregate current throughput in bits per second."
        },
        "top_talkers": {
          "items": {
            "$ref": "#/definitions/Host"
          },
          "type": "array",
          "title": "Top Talkers",
          "description": "Busiest hosts right now, by throughput."
        },
        "top_protocols": {
          "items": {
            "$ref": "#/definitions/Protocol"
          },
          "type": "array",
          "title": "Top Protocols",
          "description": "Heaviest application protocols right now."
        }
      },
      "type": "object",
      "required": [
        "summary",
        "active_hosts",
        "local_hosts",
        "active_flows",
        "throughput_bps",
        "top_talkers",
        "top_protocols"
      ],
      "title": "NetworkSummary"
    },
    "Protocol": {
      "properties": {
        "name": {
          "type": "string",
          "title": "Name",
          "description": "Application protocol name as detected by nDPI."
        },
        "bytes_total": {
          "type": "integer",
          "title": "Bytes Total",
          "description": "Bytes attributed to this protocol."
        },
        "share_percent": {
          "type": "number",
          "title": "Share Percent",
          "description": "Share of total traffic, in percent."
        }
      },
      "type": "object",
      "required": [
        "name",
        "bytes_total",
        "share_percent"
      ],
      "title": "Protocol"
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}